requirements.txt
cachetools==5.4.0
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
fonttools==4.53.1
grpcio==1.65.1
joblib==1.3.2
kiwisolver==1.4.5
loguru==0.7.2
matplotlib==3.9.1
numpy==1.26.4
numpy-ext==0.9.9
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b104
tzdata==2024.1
